+++ bindings/node/lib/bindings/post-processors.test.ts
@@ -1,0 +1,16 @@
+import { byteLevelProcessing } from "./post-processors";
+
+describe("byteLevelProcessing", () => {
+  it("instantiates correctly without any parameter", () => {
+    const processor = byteLevelProcessing();
+    expect(processor.constructor.name).toEqual("Processor");
+  });
+
+  it("accepts `undefined` as first parameter", () => {
+    expect(byteLevelProcessing(undefined)).toBeDefined();
+  });
+
+  it("accepts `boolean` as first parameter", () => {
+    expect(byteLevelProcessing(true)).toBeDefined();
+  });
+});