+++ bindings/node/lib/bindings/pre-tokenizers.test.ts
@@ -1,12 +1,18 @@
 import { byteLevelPreTokenizer, metaspacePreTokenizer } from "./pre-tokenizers";
 
 describe("byteLevelPreTokenizer", () => {
-  it("accepts `undefined` as parameter", () => {
-    expect(byteLevelPreTokenizer(undefined)).toBeDefined();
+  it("instantiates correctly", () => {
+    const processor = byteLevelPreTokenizer();
+    expect(processor.constructor.name).toEqual("PreTokenizer");
   });
 });
 
 describe("metaspacePreTokenizer", () => {
+  it("instantiates correctly without any parameter", () => {
+    const processor = metaspacePreTokenizer();
+    expect(processor.constructor.name).toEqual("PreTokenizer");
+  });
+
   it("accepts `undefined` as first parameter", () => {
     expect(metaspacePreTokenizer(undefined)).toBeDefined();
   });