+++ bindings/python/tokenizers/processors/__init__.py
@@ -3,3 +3,4 @@ from .. import processors
 PostProcessor = processors.PostProcessor
 BertProcessing = processors.BertProcessing
 RobertaProcessing = processors.RobertaProcessing
+ByteLevel = processors.ByteLevel