+++ tokenizers/src/processors/mod.rs
@@ -1,2 +1,5 @@
 pub mod bert;
 pub mod roberta;
+
+// Re-export these as processors
+pub use super::pre_tokenizers::byte_level;